[Data] Add default data provider file generator.
