+++ src/Valkyrja/Cli/Routing/Provider/ServiceProvider.php
@@ -31,8 +31,8 @@ use Valkyrja\Cli\Routing\Collector\Contract\CollectorContract;
 use Valkyrja\Cli\Routing\Data\Data;
 use Valkyrja\Cli\Routing\Dispatcher\Contract\RouterContract;
 use Valkyrja\Cli\Routing\Dispatcher\Router;
-use Valkyrja\Cli\Routing\Generator\Contract\DataFileGeneratorContract;
-use Valkyrja\Cli\Routing\Generator\DataFileGenerator;
+use Valkyrja\Cli\Routing\Generator\Contract\DataProviderFileGeneratorContract;
+use Valkyrja\Cli\Routing\Generator\DataProviderFileGenerator;
 use Valkyrja\Cli\Routing\Provider\Contract\ProviderContract;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
 use Valkyrja\Container\Provider\Provider;
@@ -48,11 +48,11 @@ final class ServiceProvider extends Provider
     public static function publishers(): array
     {
         return [
-            CollectorContract::class         => [self::class, 'publishAttributeCollector'],
-            RouterContract::class            => [self::class, 'publishRouter'],
-            CollectionContract::class        => [self::class, 'publishCollection'],
-            DataFileGeneratorContract::class => [self::class, 'publishDataFileGenerator'],
-            Data::class                      => [self::class, 'publishData'],
+            CollectorContract::class                 => [self::class, 'publishAttributeCollector'],
+            RouterContract::class                    => [self::class, 'publishRouter'],
+            CollectionContract::class                => [self::class, 'publishCollection'],
+            DataProviderFileGeneratorContract::class => [self::class, 'publishDataFileGenerator'],
+            Data::class                              => [self::class, 'publishData'],
         ];
     }
 
@@ -66,7 +66,7 @@ final class ServiceProvider extends Provider
             CollectorContract::class,
             RouterContract::class,
             CollectionContract::class,
-            DataFileGeneratorContract::class,
+            DataProviderFileGeneratorContract::class,
             Data::class,
         ];
     }
@@ -155,8 +155,8 @@ final class ServiceProvider extends Provider
         $collection = $container->getSingleton(CollectionContract::class);
 
         $container->setSingleton(
-            DataFileGeneratorContract::class,
-            new DataFileGenerator(
+            DataProviderFileGeneratorContract::class,
+            new DataProviderFileGenerator(
                 directory: $directory,
                 data: $collection->getData(),
                 namespace: $namespace,
@@ -200,7 +200,7 @@ final class ServiceProvider extends Provider
         );
         $collection->add(...$routes);
 
-        $dataGenerator = $container->getSingleton(DataFileGeneratorContract::class);
+        $dataGenerator = $container->getSingleton(DataProviderFileGeneratorContract::class);
         $dataGenerator->generateFile();
 
         $container->setSingleton(Data::class, $collection->getData());
