+++ tests/Tests/Unit/Container/Generator/DataProviderFileGeneratorTest.php
@@ -1,0 +1,155 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Valkyrja Framework package.
+ *
+ * (c) Melech Mizrachi <melechmizrachi@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Valkyrja\Tests\Unit\Container\Generator;
+
+use Valkyrja\Application\Directory\Directory;
+use Valkyrja\Container\Data\Data;
+use Valkyrja\Container\Generator\DataProviderFileGenerator;
+use Valkyrja\Support\Generator\Enum\GenerateStatus;
+use Valkyrja\Tests\EnvClass;
+use Valkyrja\Tests\Unit\Abstract\TestCase;
+
+final class DataProviderFileGeneratorTest extends TestCase
+{
+    public function testGenerateFile(): void
+    {
+        Directory::$basePath = EnvClass::APP_DIR;
+
+        $directory  = Directory::storagePath();
+        $className  = 'ContainerDataTestDataProvider';
+        $filePath   = "$directory/$className.php";
+        $data       = new Data();
+        $generator  = new DataProviderFileGenerator(
+            directory: $directory,
+            data: $data,
+            namespace: EnvClass::APP_DATA_NAMESPACE,
+            className: $className
+        );
+        $results   = $generator->generateFile();
+
+        self::assertSame(GenerateStatus::SUCCESS, $results);
+        self::assertSame($generator->generateFileContents(), @file_get_contents($filePath));
+
+        @unlink($filePath);
+    }
+
+    public function testGenerateFileContents(): void
+    {
+        $directory  = Directory::storagePath();
+        $className  = 'ContainerDataTestDataProvider';
+        $data       = new Data();
+        $generator  = new DataProviderFileGenerator(
+            directory: $directory,
+            data: $data,
+            namespace: EnvClass::APP_DATA_NAMESPACE,
+            className: $className
+        );
+        $contents  = $generator->generateFileContents();
+
+        $dataContents = $generator->generateClassContents();
+
+        $expected = <<<PHP
+            <?php
+
+            declare(strict_types=1);
+
+            // This file was automatically generated by the Valkyrja Framework.
+
+            namespace App\Provider\Data;
+
+            use Override;
+            use Valkyrja\Container\Data\Data;
+            use Valkyrja\Container\Provider\Provider;
+            use Valkyrja\Container\Manager\Contract\ContainerContract;
+
+
+            final class ContainerDataTestDataProvider extends Provider
+            {
+                /**
+                 * @inheritDoc
+                 */
+                #[Override]
+                public static function publishers(): array
+                {
+                    return [
+                        Data::class => [self::class, 'publishData'],
+                    ];
+                }
+
+                /**
+                 * @inheritDoc
+                 */
+                #[Override]
+                public static function provides(): array
+                {
+                    return [
+                        Data::class,
+                    ];
+                }
+
+                /**
+                 * Publish the service.
+                 */
+                public static function publishData(ContainerContract \$container): void
+                {
+                    \$data = $dataContents;
+
+            \$container->setSingleton(Data::class, \$data);
+                }
+            }
+
+            PHP;
+
+        self::assertNotEmpty($contents);
+        self::assertSame($expected, $contents);
+    }
+
+    public function testGenerateClassContents(): void
+    {
+        $directory  = Directory::storagePath();
+        $className  = 'ContainerDataTestDataProvider';
+        $data       = new Data();
+        $generator  = new DataProviderFileGenerator(
+            directory: $directory,
+            data: $data,
+            namespace: EnvClass::APP_DATA_NAMESPACE,
+            className: $className
+        );
+        $contents  = $generator->generateClassContents();
+
+        $dataNamespace = Data::class;
+
+        // phpcs:disable
+        $expected = <<<PHP
+            new \\$dataNamespace(
+                aliases: array (
+            ),
+                deferred: array (
+            ),
+                deferredCallback: array (
+            ),
+                services: array (
+            ),
+                singletons: array (
+            ),
+                providers: array (
+            )
+            )
+            PHP;
+        // phpcs:enable
+
+        self::assertNotEmpty($contents);
+        self::assertSame($expected, $contents);
+    }
+}
