+++ tests/Tests/Unit/Http/Routing/Generator/DefaultDataProviderFileGeneratorTest.php
@@ -1,0 +1,108 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Valkyrja Framework package.
+ *
+ * (c) Melech Mizrachi <melechmizrachi@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Valkyrja\Tests\Unit\Http\Routing\Generator;
+
+use Valkyrja\Application\Directory\Directory;
+use Valkyrja\Http\Routing\Generator\DefaultDataProviderFileGenerator;
+use Valkyrja\Support\Generator\Enum\GenerateStatus;
+use Valkyrja\Tests\EnvClass;
+use Valkyrja\Tests\Unit\Abstract\TestCase;
+
+final class DefaultDataProviderFileGeneratorTest extends TestCase
+{
+    public function testGenerateFile(): void
+    {
+        Directory::$basePath = EnvClass::APP_DIR;
+
+        $directory  = Directory::storagePath();
+        $className  = 'HttpRoutingDataTestDataProvider';
+        $filePath   = "$directory/$className.php";
+        $generator  = new DefaultDataProviderFileGenerator(
+            directory: $directory,
+            namespace: EnvClass::APP_DATA_NAMESPACE,
+            className: $className
+        );
+        $results   = $generator->generateFile();
+
+        self::assertSame(GenerateStatus::SUCCESS, $results);
+        self::assertSame($generator->generateFileContents(), @file_get_contents($filePath));
+
+        @unlink($filePath);
+    }
+
+    public function testGenerateFileContents(): void
+    {
+        $directory  = Directory::storagePath();
+        $className  = 'HttpRoutingDataTestDataProvider';
+        $generator  = new DefaultDataProviderFileGenerator(
+            directory: $directory,
+            namespace: EnvClass::APP_DATA_NAMESPACE,
+            className: $className
+        );
+        $contents  = $generator->generateFileContents();
+
+        $expected = <<<PHP
+            <?php
+
+            declare(strict_types=1);
+
+            // This file was automatically generated by the Valkyrja Framework.
+
+            namespace App\Provider\Data;
+
+            use Override;
+            use Valkyrja\Http\Routing\Data\Data;
+            use Valkyrja\Container\Provider\Provider;
+            use Valkyrja\Container\Manager\Contract\ContainerContract;
+            use Valkyrja\Http\Routing\Provider\ServiceProvider;
+
+            final class HttpRoutingDataTestDataProvider extends Provider
+            {
+                /**
+                 * @inheritDoc
+                 */
+                #[Override]
+                public static function publishers(): array
+                {
+                    return [
+                        Data::class => [self::class, 'publishData'],
+                    ];
+                }
+
+                /**
+                 * @inheritDoc
+                 */
+                #[Override]
+                public static function provides(): array
+                {
+                    return [
+                        Data::class,
+                    ];
+                }
+
+                /**
+                 * Publish the service.
+                 */
+                public static function publishData(ContainerContract \$container): void
+                {
+                    ServiceProvider::publishData(\$container);
+                }
+            }
+
+            PHP;
+
+        self::assertNotEmpty($contents);
+        self::assertSame($expected, $contents);
+    }
+}
